requirements.txt
# Core dependencies
langgraph>=0.2.0
openai>=1.0.0
requests>=2.31.0
python-dotenv>=1.0.0

# Vector database and embeddings
chromadb>=0.4.0
sentence-transformers>=2.2.0

# PDF processing
PyPDF2>=3.0.0

# Additional utilities
typing-extensions>=4.0.0
pydantic>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0

# Optional: Web framework for API
fastapi>=0.104.0
uvicorn>=0.24.0